pyproject.toml
[project]
name = "reforgie"
version = "0.1.0"
description = "An alternative to ModBuddy for building simple, streamlined Civ V mods."
authors = [
    {name = "AcipenserSturio", email = "AcipenserSturio@users.noreply.github.com"},
]
dependencies = [
    "Pillow>=10.0.1",
    "tomli>=2.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "GPL-3.0"}
